+++ spec/factories/organizations.rb
@@ -3,5 +3,6 @@
 FactoryBot.define do
   factory :organization do
     sequence(:organization_name) { |n| "#{Faker::Company.name}-#{n}" }
+    sequence(:mlid) { |n| n.to_s(36).rjust(3, '0').upcase }
   end
 end